feat: restructure backend application and add initial configuration files
