+++ apps/backend/.env.dev
@@ -1,0 +1,8 @@
+DATABASE_URL=your_database_url_here
+
+# JWT TOKEN CONFIGURATION
+
+JWT_SECRET=your_jwt_secret_here
+JWT_TOKEN_AUDIENCE=your_jwt_audience_here
+JWT_TOKEN_ISSUER=your_jwt_issuer_here
+JWT_TTL="30d"
